+++ NEWS.md
@@ -1,3 +1,12 @@
+# bayfox v0.0.1a4
+
+## Enhancements
+
+
+## Bug fixes
+
+
+
 # bayfox v0.0.1a3
 
 ## Enhancements